Decouple AuthProvider from firebase - rename, use own User object
