+++ src/components/SignUp/SignUp.tsx
@@ -4,7 +4,7 @@ import firebase from "firebase/app";
 import "firebase/auth";
 import "firebase/firestore";
 
-import { AuthContext } from "../../Firebase/FireBaseAuthProvider";
+import { AuthContext } from "../AuthProvider/AuthProvider";
 
 interface FormItems {
   username: string;
